Factor out `toTree` into backend.
Summary:
We move the `toTree` method of our intermediate representation in the
type generator to the backend. The method is replace by a very simple
visitor.
