+++ type-generator/src/main/scala/mesosphere/raml/backend/treehugger/StringVisitor.scala
@@ -1,0 +1,10 @@
+package mesosphere.raml.backend.treehugger
+
+import mesosphere.raml.ir.StringT
+
+import treehugger.forest.Tree
+
+object StringVisitor {
+
+  def visit(s: StringT): Seq[Tree] =  Seq.empty[Tree]
+}
